[FIX] favorites, pay 엔티티 및 관련 기능 수정

+++ src/main/java/com/example/parking/repository/PayRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PayRepository extends JpaRepository<Pay, Integer> {
-    List<Pay> findByUserUserId(Integer userId);
+public interface PayRepository extends JpaRepository<Pay, Long> {
+    List<Pay> findByUserUserId(Long userId);
 }
